Improve maintainability of website (#94)
* Restructure to make site easier to maintain

* Seperate website and documentation

* Configure website part for new directory structure

* Update docfx config to genereate articles

* Hook docfx generation into eleventy events

* Created three separate configs for docfx based on all, api, or just articles for dev

* Removed files that are no longer needed

* Added main navigation toc for docfx pages

* Updated to include the .md extensions like it originally did

* Moved foundation page into documentation since it's pure md file

* Presskit uses base layout now

* Create monogame template for docfx

* New npm scripts to use when developing to specify website, api, or articles only for generation. This is to help speed up development hotreloading when reloading all api documents isn't neccessary.

* Ignore all of the website directory if we are only working on api or articles.

* Fixed infinite loop bug where watching docfx documenation files would trigger a build, which triggers the watch to rebuild which triggers docfx to rebuild, etc, etc.

* Move docfxBuilder to its own module file.

* Remove index.js

This was being used to processes the env variable to determien which configuration of docfx to use.  The environment variable logic was moved to the script in package.json using the %npm_lifecycle_event% variable to determien the run mode to use.

* Renamed to docfx.js

* Created docfx.js

This was initially docfxBuilder, which was renamed to docfx.  Outputs relavent information into console when building to inform the the user which docfx files are being generated (articles, api, both, or neither).

* Update path to docfx after file rename.

* Output newline to console

Ensure that a newline is output to console when finished to seperate docfx output from 11ty output.

* Add sitemap generation.

* Remove unneeded excludes

* _appLogoPath not used

Intead this is used directly in the template to pick the logo path up from the 11ty output

* Update year in footer

* Simplify scripts and ensure cross os compatible

The scripts were simplified from 12 scripts to 4.  Setting the RUN_MODE environment variable is also done in a cross platform compatbiel way between Windoes, Mac, and Linux system.

Using the cross-env package was considered but initial testing showed issues where it would spawn multiple processes due to the terminal signal chaining, which could cause the termianl to hang after ending the initial process.

* Don't output logo and favicon from docfx

11ty generates the logo and favicon files, docfx should not also output these as they will be overwritten anyway.

* Use `<xref>` for API document links

DocFx will internally resolve `<xref>` links appropriatly to the API documentation.

* Perform dotnet tool restore

This ensures that the docfx tool is restored before executing the command.  This way users dont' have to worry about manually restoring it themselves.

* Update readme

Readme updated to reflect changes in the new npm script commands.

* cross-env is needed after testing to ensure compatibility between OSs

* Update scripts to use cross-env

* Add twitter and rss icons to documentation footer

* Use same theme switch icons between site and documentation

* No longer need to run docfx as separate command in workflow

* Remove excludes from configuration

The excludes are not needed and can be removed to simplify the configuration file(s).

* Clean up resources confguration

There is no codeCnippit or scripts directory to pull resources from. Removed these to simplify configuration file(s).

* Include the foundation directory in build for all

Ensure the foundation .md files are built as part of the all configuration.

* Use correct opengraph image path.

* Change hostname from monogame.github.io to monogame.net

* Update year in footer

* Use correct path for favicon

* Update favicon path in template

Template was using the _rel path which is incorrect. The path should not be relative.

* Update to ensure api/index.md is not ignored.

* Fix broken links on documentation page.