Fix MacOS net8_mgfxc_wine_setup.sh (#145)
Some issues with the install script caused errors on macos. 
The default macos `grep` does not support the -P perl flag. 
So use a basic regex instead.

We are trying to run `wine` in one part of the script, which will fail on M* Apple Macs because it is a 32bit executable. 

Also upgade to .net 8.0.401